doc, tools: extract linkJsTypeDocs
