+++ tools/doc/html.js
@@ -15,6 +15,7 @@ const toID = require('./lib/toID')
 const render = require('./lib/render')
 const parseLists = require('./lib/parseLists')
 const parseYAML = require('./lib/parseYAML')
+const linkJsTypeDocs = require('./lib/linkJsTypeDocs')
 
 module.exports = toHTML;
 
@@ -28,27 +29,6 @@ marked.setOptions({
 });
 
 
-
-
-function linkJsTypeDocs(text) {
-  var parts = text.split('`');
-  var i;
-  var typeMatches;
-
-  // Handle types, for example the source Markdown might say
-  // "This argument should be a {Number} or {String}"
-  for (i = 0; i < parts.length; i += 2) {
-    typeMatches = parts[i].match(/\{([^\}]+)\}/g);
-    if (typeMatches) {
-      typeMatches.forEach(function(typeMatch) {
-        parts[i] = parts[i].replace(typeMatch, typeParser.toLink(typeMatch));
-      });
-    }
-  }
-
-  //XXX maybe put more stuff here?
-  return parts.join('`');
-}
 
 function parseAPIHeader(text) {
   text = text.replace(
