+++ tools/doc/lib/linkJsTypeDocs.js
@@ -1,0 +1,19 @@
+module.exports = function linkJsTypeDocs(text) {
+  var parts = text.split('`');
+  var i;
+  var typeMatches;
+
+  // Handle types, for example the source Markdown might say
+  // "This argument should be a {Number} or {String}"
+  for (i = 0; i < parts.length; i += 2) {
+    typeMatches = parts[i].match(/\{([^\}]+)\}/g);
+    if (typeMatches) {
+      typeMatches.forEach(function(typeMatch) {
+        parts[i] = parts[i].replace(typeMatch, typeParser.toLink(typeMatch));
+      });
+    }
+  }
+
+  //XXX maybe put more stuff here?
+  return parts.join('`');
+}
